+++ src/RcppExports.cpp
@@ -156,14 +156,14 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// reverse_index_founders
-IntegerVector reverse_index_founders(IntegerVector cross_info);
-RcppExport SEXP qtl2geno_reverse_index_founders(SEXP cross_infoSEXP) {
+// invert_founder_index
+IntegerVector invert_founder_index(IntegerVector cross_info);
+RcppExport SEXP qtl2geno_invert_founder_index(SEXP cross_infoSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< IntegerVector >::type cross_info(cross_infoSEXP);
-    rcpp_result_gen = Rcpp::wrap(reverse_index_founders(cross_info));
+    rcpp_result_gen = Rcpp::wrap(invert_founder_index(cross_info));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -304,6 +304,29 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const double >::type tol(tolSEXP);
     Rcpp::traits::input_parameter< const bool >::type verbose(verboseSEXP);
     rcpp_result_gen = Rcpp::wrap(est_map(crosstype, genotypes, founder_geno, is_X_chr, is_female, cross_info, rec_frac, error_prob, max_iterations, tol, verbose));
+    return rcpp_result_gen;
+END_RCPP
+}
+// est_map2
+NumericVector est_map2(const String& crosstype, const IntegerMatrix& genotypes, const IntegerMatrix& founder_geno, const bool is_X_chr, const LogicalVector& is_female, const IntegerMatrix& cross_info, const IntegerVector& cross_group, const IntegerVector& unique_cross_group, const NumericVector& rec_frac, const double error_prob, const int max_iterations, const double tol, const bool verbose);
+RcppExport SEXP qtl2geno_est_map2(SEXP crosstypeSEXP, SEXP genotypesSEXP, SEXP founder_genoSEXP, SEXP is_X_chrSEXP, SEXP is_femaleSEXP, SEXP cross_infoSEXP, SEXP cross_groupSEXP, SEXP unique_cross_groupSEXP, SEXP rec_fracSEXP, SEXP error_probSEXP, SEXP max_iterationsSEXP, SEXP tolSEXP, SEXP verboseSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const String& >::type crosstype(crosstypeSEXP);
+    Rcpp::traits::input_parameter< const IntegerMatrix& >::type genotypes(genotypesSEXP);
+    Rcpp::traits::input_parameter< const IntegerMatrix& >::type founder_geno(founder_genoSEXP);
+    Rcpp::traits::input_parameter< const bool >::type is_X_chr(is_X_chrSEXP);
+    Rcpp::traits::input_parameter< const LogicalVector& >::type is_female(is_femaleSEXP);
+    Rcpp::traits::input_parameter< const IntegerMatrix& >::type cross_info(cross_infoSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector& >::type cross_group(cross_groupSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector& >::type unique_cross_group(unique_cross_groupSEXP);
+    Rcpp::traits::input_parameter< const NumericVector& >::type rec_frac(rec_fracSEXP);
+    Rcpp::traits::input_parameter< const double >::type error_prob(error_probSEXP);
+    Rcpp::traits::input_parameter< const int >::type max_iterations(max_iterationsSEXP);
+    Rcpp::traits::input_parameter< const double >::type tol(tolSEXP);
+    Rcpp::traits::input_parameter< const bool >::type verbose(verboseSEXP);
+    rcpp_result_gen = Rcpp::wrap(est_map2(crosstype, genotypes, founder_geno, is_X_chr, is_female, cross_info, cross_group, unique_cross_group, rec_frac, error_prob, max_iterations, tol, verbose));
     return rcpp_result_gen;
 END_RCPP
 }