+++ src/cross_dof1.cpp
@@ -359,3 +359,26 @@ const double DOF1::est_rec_frac(const NumericVector& gamma, const bool is_x_chr,
 
     return NA_REAL;
 }
+
+// not implemented
+const NumericVector DOF1::est_map2(const IntegerMatrix& genotypes,
+                                   const IntegerMatrix& founder_geno,
+                                   const bool is_X_chr,
+                                   const LogicalVector& is_female,
+                                   const IntegerMatrix& cross_info,
+                                   const IntegerVector& cross_group,
+                                   const IntegerVector& unique_cross_group,
+                                   const NumericVector& rec_frac,
+                                   const double error_prob,
+                                   const int max_iterations,
+                                   const double tol,
+                                   const bool verbose)
+{
+    Rcpp::stop("est_map not yet implemented for DO F1s.");
+
+    // return vector of NAs
+    const int n_rf = rec_frac.size();
+    NumericVector result(n_rf);
+    for(int i=0; i<n_rf; i++) result[i] = NA_REAL;
+    return result ;
+}