+++ src/cross_dof1.h
@@ -45,6 +45,21 @@ class DOF1 : public QTLCross
 
     const double est_rec_frac(const Rcpp::NumericVector& gamma, const bool is_x_chr,
                               const Rcpp::IntegerMatrix& cross_info, const int n_gen);
+
+    // not yet implemented
+    const Rcpp::NumericVector est_map2(const Rcpp::IntegerMatrix& genotypes,
+                                       const Rcpp::IntegerMatrix& founder_geno,
+                                       const bool is_X_chr,
+                                       const Rcpp::LogicalVector& is_female,
+                                       const Rcpp::IntegerMatrix& cross_info,
+                                       const Rcpp::IntegerVector& cross_group,
+                                       const Rcpp::IntegerVector& unique_cross_group,
+                                       const Rcpp::NumericVector& rec_frac,
+                                       const double error_prob,
+                                       const int max_iterations,
+                                       const double tol,
+                                       const bool verbose);
+
 };
 
 #endif // CROSS_DOF1_H