+++ src/cross_hs.h
@@ -49,6 +49,21 @@ class HS : public QTLCross
     const int nrec(const int gen_left, const int gen_right,
                    const bool is_x_chr, const bool is_female,
                    const Rcpp::IntegerVector& cross_info);
+
+    // not yet implemented
+    const Rcpp::NumericVector est_map2(const Rcpp::IntegerMatrix& genotypes,
+                                       const Rcpp::IntegerMatrix& founder_geno,
+                                       const bool is_X_chr,
+                                       const Rcpp::LogicalVector& is_female,
+                                       const Rcpp::IntegerMatrix& cross_info,
+                                       const Rcpp::IntegerVector& cross_group,
+                                       const Rcpp::IntegerVector& unique_cross_group,
+                                       const Rcpp::NumericVector& rec_frac,
+                                       const double error_prob,
+                                       const int max_iterations,
+                                       const double tol,
+                                       const bool verbose);
+
 };
 
 #endif // CROSS_HS_H