+++ src/cross_riself8.h
@@ -49,6 +49,20 @@ class RISELF8 : public QTLCross
     // check whether X chr can be handled
     const bool check_handle_x_chr(const bool any_x_chr);
 
+    // tailored est_map that pre-calculates transition matrices, etc
+    const Rcpp::NumericVector est_map2(const Rcpp::IntegerMatrix& genotypes,
+                                       const Rcpp::IntegerMatrix& founder_geno,
+                                       const bool is_X_chr,
+                                       const Rcpp::LogicalVector& is_female,
+                                       const Rcpp::IntegerMatrix& cross_info,
+                                       const Rcpp::IntegerVector& cross_group,
+                                       const Rcpp::IntegerVector& unique_cross_group,
+                                       const Rcpp::NumericVector& rec_frac,
+                                       const double error_prob,
+                                       const int max_iterations,
+                                       const double tol,
+                                       const bool verbose);
+
 };
 
 #endif // CROSS_RISELF8_H